update, delete data

+++ app/Http/Controllers/UserController.php
@@ -4,30 +4,26 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
     public function index()
     {
-        // $users = User::where('id', '!=', 1)->orderBy('name')->paginate(10);
-        // return view('user.index', compact('users'));
-
         $search = request('search');
 
         if ($search) {
             $users = User::where(function ($query) use ($search) {
-                $query->where('name', 'like', '%' . $search . '%') // Perbaikan di sini
-                      ->orWhere('email', 'like', '%' . $search . '%'); // Perbaikan di sini
+                $query->where('name', 'like', '%' . $search . '%')
+                    ->orWhere('email', 'like', '%' . $search . '%');
             })
-            ->orderBy('name')
-            ->where('id', '!=', 1)
-            ->paginate(20)
-            ->withQueryString();
+                ->orderBy('name')
+                ->where('id', '!=', 1)
+                ->paginate(20)
+                ->withQueryString();
         } else {
             $users = User::where('id', '!=', 1)
-            ->orderBy('name')
-            ->paginate(10);
+                ->orderBy('name')
+                ->paginate(10);
         }
 
         return view('user.index', compact('users'));
@@ -35,12 +31,33 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
-        $user->delete(); // Menghapus user
-        return redirect()->route('user.index')->with('success', 'User  deleted successfully!');
+        if ($user->id != 1) {
+            $user->delete();
+            return back()->with('success', 'delete user successfully!');
+        } else {
+            return redirect()->route('user.index')->with('danger', 'Delete user failed!');
+        }
     }
 
-    public function edit(User $user)
+    public function makeadmin(User $user)
     {
-        return view('user.edit', compact('user')); // Mengirimkan user ke view edit
+        $user->timestamps = false;
+        $user->is_admin = true;
+        $user->save();
+
+        return back()->with('success', 'Make Admin Successfully!');
+    }
+
+    public function removeadmin(User $user)
+    {
+        if ($user->id != 1) {
+            $user->timestamps = false;
+            $user->is_admin = false;
+            $user->save();
+
+            return back()->with('success', 'Remove Admin Successfully!');
+        } else {
+            return redirect()->route('user.index');
+        }
     }
 }
